Fixed UI bugs
