+++ src/components/ViewCard/Comments.tsx
@@ -1,15 +1,11 @@
 import size from "lodash/size";
 import styled from "styled-components";
-import { faUser, faPlus } from "@fortawesome/free-solid-svg-icons";
+import { faUser } from "@fortawesome/free-solid-svg-icons";
 import { Avatar } from "@deskpro/deskpro-ui";
 import ReactTimeAgo from "react-time-ago";
-import {
-    H1,
-    P1,
-    P11,
-    Stack,
-    Button as ButtonUI,
-} from "@deskpro/app-sdk";
+import { P1, P11, Stack } from "@deskpro/deskpro-ui";
+import { Title } from "@deskpro/app-sdk";
+import { Container } from "../common";
 import type { FC } from "react";
 import type { Comment } from "../../services/trello/types";
 import type { Maybe } from "../../types";
@@ -30,17 +26,11 @@ const Comments: FC<{
 }> = ({ comments, onClickTitleAction }) => {
     return (Array.isArray(comments) && size(comments))
         ? (
-            <>
-                <Stack gap={6} align="center" style={{ marginBottom: 14 }}>
-                    <H1>Comments  ({comments.length})</H1>
-                    <ButtonUI
-                        icon={faPlus}
-                        minimal
-                        noMinimalUnderline
-                        onClick={onClickTitleAction}
-                    />
-                </Stack>
-
+            <Container>
+                <Title
+                    title={`Comments  (${size(comments)})`}
+                    onClick={onClickTitleAction}
+                />
                 {comments.map(({ id, date, data, memberCreator }) => (
                     <Stack key={id} wrap="nowrap" gap={6} style={{ marginBottom: 10 }}>
                         <Stack vertical>
@@ -52,7 +42,7 @@ const Comments: FC<{
                         <CommentBlock>{data.text}</CommentBlock>
                     </Stack>
                 ))}
-            </>
+            </Container>
         )
         : null;
 }
