+++ src/components/ViewCard/Info.tsx
@@ -1,0 +1,92 @@
+import get from "lodash/get";
+import styled from "styled-components";
+import { faCalendarDays } from "@fortawesome/free-solid-svg-icons";
+import { Title, useDeskproAppTheme } from "@deskpro/app-sdk";
+import { P5, Pill, Icon, Stack } from "@deskpro/deskpro-ui";
+import { getDate } from "../../utils/date";
+import { getLabelColor } from "../../utils";
+import {
+    LinkIcon,
+    Container,
+    TrelloLogo,
+    EmptyInlineBlock,
+    TextBlockWithLabel,
+} from "../common";
+import { Members } from "../common/Cards";
+import type { FC } from "react";
+import type { CardType } from "../../services/trello/types";
+
+type Props = {
+    card: CardType,
+};
+
+const Description = styled(P5)`
+    white-space: pre-wrap
+`;
+
+const Info: FC<Props> = ({ card }) => {
+    const { theme } = useDeskproAppTheme();
+    const labels = get(card, ["labels"]);
+    const due = get(card, ["due"]);
+
+    return (
+        <Container>
+            <Title
+                title={get(card, ["name"], "-")}
+                icon={<TrelloLogo/>}
+                link={get(card, ["shortUrl"], "#")}
+            />
+            <TextBlockWithLabel
+                label="Board"
+                text={(
+                    <>
+                        <P5 style={{ marginRight: 4 }}>{get(card, ["board", "name"], "-")}</P5>
+                        <LinkIcon size={10} href={get(card, ["board", "url"], "-")}/>
+                    </>
+                )}
+            />
+            <TextBlockWithLabel
+                label="List"
+                text={get(card, ["list", "name"], "-")}
+            />
+            <TextBlockWithLabel
+                label="Description"
+                text={<Description>{get(card, ["desc"], "-")}</Description>}
+            />
+            <TextBlockWithLabel
+                label="Labels"
+                text={(!labels)
+                    ? "-"
+                    : (
+                        <Stack wrap="wrap" gap={6}>
+                            {
+                                // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                                labels.map(({ id, name, color }: any) => (
+                                    <Pill
+                                        key={id}
+                                        label={name ? name : <EmptyInlineBlock/>}
+                                        {...getLabelColor(theme, color)}
+                                    />
+                                ))
+                            }
+                        </Stack>
+                    )
+                }
+            />
+            <TextBlockWithLabel
+                label="Due date"
+                text={(
+                    <P5>
+                        {/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
+                        {/* @ts-ignore */}
+                        <Icon icon={faCalendarDays} style={{ color: theme.colors.grey40 }}/>&nbsp;
+                        {!due ? "-" : getDate(due)}
+                    </P5>
+                )}
+            />
+            <Members members={get(card, ["members"])} />
+        </Container>
+    );
+};
+
+export { Info };
