+++ src/components/common/TrelloLink/index.ts
@@ -1,1 +1,2 @@
 export { TrelloLink } from "./TrelloLink";
+export { TrelloLogo } from "./TrelloLogo";
